+++ test_empty_prompt.py
@@ -1,5 +1,3 @@
-from unittest.mock import AsyncMock
 
-import pytest
 
 # Will be implemented